refactor(sentry): extract the autofix family resolver from select.mjs
`scripts/sentry-autofix-select.mjs` sat at 737 lines, over the recurring-review
checklist's 600-line soft cap. Move the live-state family-resolution
orchestration (`resolveFamilies` + the MAX_REVERSE_ITERATIONS fixpoint ceiling)
into a new `scripts/sentry-autofix-family-resolve.mjs`. The selector keeps the
pure parse / evaluate / report / CLI layers and calls the resolver the same way;
`sentry-autofix-family.mjs` stays the pure union-find, and the new module is the
`gh`-driven resolver on top of it.

Behavior is unchanged — `resolveFamilies` was already internal (no importer
outside select.mjs), so no public export moved; its coverage stays in the select
suite, which drives it end to end through `selectAutofixRun`. select.mjs is now
565 lines. Route the new module to the select suite in the quality gate.
